Create webpage with header and footer, input, and buttons
Add page title, add header and footer content, add initial input value to
instruct user, add two buttons for creation of the etch-a-sketch board
and removal of board elements, respectively. In this version the create
button works but the reset button function has yet to be added.
Board is not yet functional, just a display.