Change 'mystery' to standalone function, see desc.
'mystery' now runs when user clicks the associated button.
setInterval is now used to call colorPixel, which colors
each child of 'container' using a forEach hack on the Nodelist,
via colorChanger (which brushTool also uses). The anonymous
function inside mystery's onclick event calls setInterval,
which in turn calls colorPixel with a counter, which is then
passed to colorChanger as a modifier on the three brushCounters,
increasing the rate of each oscillator linearly. This produces
an evolving trippy phasing effect on the board. Other functions
may also be applied in the call to colorPixel.